style: type alias for ptrdiff_t
added a type alias for `ptrdiff_t` for legibility.

+++ internal/mi/mi_application.go
@@ -8,7 +8,7 @@ import (
 // _MI_Application
 type Application struct {
 	reserved1 uint64
-	reserved2 int64
+	reserved2 ptrdiff_t
 	ft        *applicationFT
 }
